Remove hardcoded items from the tasks array, Use js modules
